Updated the structure and fixed small issues
