+++ gestalt/helpers.py
