--- gestalt/provider/__init__.py
@@ -1,1 +1,0 @@
-import provider
