+++ tests/testvault/testjson.yaml
@@ -1,0 +1,3 @@
+{
+  "test_mount": "ref+vault://test-mount/data/test"
+}
